github/workflows: specify golang version as environment variable.
using a matrix variant creates another dimension which needs to
be addressed in matrix includes (the linux/arm64 include would
need the go compiler specified as well, which would make the
value bleed into too many locations). since action parameters are
translated into environment variables, we can specify the variable
globally. (alternatively we could also specify it on the job
step, but this is too far down the file and requires searching)